Split the Hid code up in a few files to make it more practical to work on
