+++ Windows/Hid/DualSense.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "Common/CommonWindows.h"
+
+bool InitializeDualSense(HANDLE handle, int outReportSize);
+bool ShutdownDualsense(HANDLE handle, int outReportSize);
+bool ReadDualSenseInput(HANDLE handle, HIDControllerState *state);
