+++ Windows/Hid/HidInputDevice.h
@@ -1,11 +1,15 @@
 // This file in particular along with its cpp file is public domain, use it for whatever you want.
 
+// Internal common header for Hid input stuff.
+
 #pragma once
 
+#include <set>
+
+#include "Common/CommonTypes.h"
 #include "Common/Input/InputState.h"
 #include "Windows/InputDevice.h"
-#include <set>
-#include <windows.h>
+#include "Common/CommonWindows.h"
 
 enum class HIDControllerType {
 	DS4,
@@ -26,7 +30,10 @@ struct HIDControllerState {
 	float accelerometer[3];  // X, Y, Z
 };
 
-struct ButtonInputMapping;
+struct ButtonInputMapping {
+	u32 button;
+	InputKeyCode keyCode;
+};
 
 // Supports a few specific HID input devices, namely DualShock and DualSense.
 // More may be added later. Just picks the first one available, for now.
@@ -44,7 +51,6 @@ private:
 	void ReleaseAllKeys(const ButtonInputMapping *buttonMappings, int count);
 	InputDeviceID DeviceID(int pad);
 	HIDControllerState prevState_{};
-
 	HIDControllerType subType_{};
 	HANDLE controller_;
 	std::string name_;
